🔧 refactor: increase string length limits for various columns in database models to improve flexibility

+++ backend/app/db/models/group_settings.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, ForeignKey
+from sqlalchemy import Column, Integer, BigInteger, ForeignKey
 from sqlalchemy.dialects.sqlite import JSON as SQLiteJSON
 from sqlalchemy.orm import relationship
 
@@ -15,7 +15,7 @@ class GroupSettings(Base):
     # “mutable default” foot‑gun when someone mutates the list in place.
     allowed_extensions = Column(SQLiteJSON, default=list)
 
-    max_file_size = Column(Integer, default=10_000_000)      # bytes
-    max_storage_size = Column(Integer, nullable=True)        # None => unlimited
+    max_file_size = Column(BigInteger, default=10_000_000)      # bytes
+    max_storage_size = Column(BigInteger, nullable=True)        # None => unlimited
 
     group = relationship("Group", back_populates="settings")
